requirements.txt
appdirs==1.4.4
asgiref==3.7.2
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
cffi
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.4
cryptography==41.0.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.3
django-allauth==0.43.0
django-crispy-forms==1.9.2
django-extensions==3.0.9
django-js-asset==1.2.2
django-taggit==1.3.0
filelock==3.0.12
idna==2.10
isort==5.6.4
joblib==1.3.1
lazy-object-proxy==1.4.3
lxml==4.9.3
mccabe==0.6.1
nltk==3.8.1
oauthlib==3.1.0
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==10.0.0
pipenv==2020.5.28
protobuf==3.13.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
PyPDF2==1.26.0
python-docx==0.8.10
python3-openid==3.2.0
pytz==2020.4
regex==2023.6.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==1.25.11
virtualenv==20.1.0
virtualenv-clone==0.5.4
Wand==0.6.11
wincertstore==0.2
wrapt==1.12.1